+++ src/main/java/ee/yals/models/dao/SecretRepo.java
@@ -1,4 +1,4 @@
-package ee.yals.models.repo;
+package ee.yals.models.dao;
 
 import ee.yals.models.Secret;
 import ee.yals.models.User;
@@ -7,7 +7,7 @@ import org.springframework.data.repository.Repository;
 import java.util.List;
 import java.util.Optional;
 
-public interface SecretRepo extends Repository<Secret, Long> {
+interface SecretRepo extends Repository<Secret, Long> {
     Optional<Secret> findSingleByUser(User user);
 
     Secret save(Secret secretToSave);
