+++ src/main/java/ee/yals/models/dao/TokenDao.java
@@ -1,7 +1,6 @@
 package ee.yals.models.dao;
 
 import ee.yals.models.Token;
-import ee.yals.models.repo.TokenRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
