+++ src/main/java/ee/yals/models/dao/TokenRepo.java
@@ -1,11 +1,11 @@
-package ee.yals.models.repo;
+package ee.yals.models.dao;
 
 import ee.yals.models.Token;
 import org.springframework.data.repository.Repository;
 
 import java.util.Optional;
 
-public interface TokenRepo extends Repository<Token, Long> {
+interface TokenRepo extends Repository<Token, Long> {
     Optional<Token> findSingleByToken(String token);
 
     Token save(Token tokenToSave);
